+++ src/languages.ts
@@ -1,12 +1,6 @@
 import * as fs from 'fs/promises';
 import * as path from 'path';
-
-import * as plugins from '../plugins';
-
-// Define an interface for the language data structure
-interface LanguageData {
-    [key: string]: string;
-}
+import * as plugins from './plugins';
 
 // Define a type for the specific error you're handling
 interface CustomError extends NodeJS.ErrnoException {
@@ -14,24 +8,21 @@ interface CustomError extends NodeJS.ErrnoException {
 }
 
 // Export the function for getting language data
-export async function get(language: string, namespace: string): Promise<LanguageData> {
+export async function get(language: string, namespace: string): Promise<Record<string, string>> {
     const languagesPath = path.join(__dirname, '../build/public/language');
     const pathToLanguageFile = path.join(languagesPath, language, `${namespace}.json`);
     if (!pathToLanguageFile.startsWith(languagesPath)) {
         throw new Error('[[error:invalid-path]]');
     }
+
     const data = await fs.readFile(pathToLanguageFile, 'utf8');
-
-
-
-
-    const parsed: LanguageData = JSON.parse(data) as LanguageData || {};
-
+    const parsed = JSON.parse(data) as Record<string, string>;
     const result = await plugins.hooks.fire('filter:languages.get', {
         language,
         namespace,
         data: parsed,
-    }) as { data: LanguageData }; // Type assertion here
+    }) as { data: Record<string, string> };
+
     return result.data;
 }
 
@@ -44,17 +35,14 @@ export async function listCodes(): Promise<string[]> {
     try {
         const languagesPath = path.join(__dirname, '../build/public/language');
         const file = await fs.readFile(path.join(languagesPath, 'metadata.json'), 'utf8');
-
-        // Manually parse the JSON string, which TypeScript won't recognize as any
         const parsed = JSON.parse(file) as { languages: string[] };
         codeCache = parsed.languages;
         return parsed.languages;
     } catch (err) {
-        const customError = err as CustomError; // Type assertion here
+        const customError = err as CustomError;
         if (customError.code === 'ENOENT') {
             return [];
         }
         throw err;
     }
 }
-
